Two missing fields in RTCM3 1001 parsing.
Turns out we can't use short in the rtcm3 structure, JSON types won't
map to it. Well, in theory we could map int to it with a range check
at parsing time, but it's not worth the effort.

All regression tests pass.

+++ gps.h
@@ -270,7 +270,7 @@ struct rtcm3_rtk_hdr {		/* header data from 1001, 1002, 1003, 1004 */
     bool sync;			/* Synchronous GNSS Message Flag */
     unsigned short satcount;	/* # Satellite Signals Processed */
     bool smoothing;		/* Divergence-free Smoothing Indicator */
-    unsigned short interval;	/* Smoothing Interval */
+    unsigned int interval;	/* Smoothing Interval */
 };
 
 struct rtcm3_basic_rtk {
